+++ tfjs-core/src/ops/ops.ts
@@ -72,13 +72,16 @@ export {exp} from './exp';
 export {expandDims} from './expand_dims';
 export {expm1} from './expm1';
 export {eye} from './eye';
+export {fft} from './fft';
 export {fill} from './fill';
 export {floor} from './floor';
 export {floorDiv} from './floorDiv';
 export {gather} from './gather';
 export {greater} from './greater';
 export {greaterEqual} from './greater_equal';
+export {ifft} from './ifft';
 export {imag} from './imag';
+export {irfft} from './irfft';
 export {leakyRelu} from './leaky_relu';
 export {less} from './less';
 export {lessEqual} from './less_equal';
@@ -138,6 +141,7 @@ export {reverse1d} from './reverse_1d';
 export {reverse2d} from './reverse_2d';
 export {reverse3d} from './reverse_3d';
 export {reverse4d} from './reverse_4d';
+export {rfft} from './rfft';
 export {scalar} from './scalar';
 export {selu} from './selu';
 export {separableConv2d} from './separable_conv2d';
@@ -188,7 +192,6 @@ export * from './transpose';
 export * from './norm';
 export * from './moving_average';
 export * from './scatter_nd';
-export * from './spectral_ops';
 export * from './sparse_to_dense';
 export * from './gather_nd';
 export * from './dropout';
@@ -197,7 +200,17 @@ export * from './in_top_k';
 
 export {op} from './operation';
 
-import * as spectral from './spectral_ops';
+import {rfft} from './rfft';
+import {fft} from './fft';
+import {ifft} from './ifft';
+import {irfft} from './irfft';
+const spectral = {
+  fft,
+  ifft,
+  rfft,
+  irfft
+};
+
 import * as fused from './fused_ops';
 
 import {hammingWindow} from './hamming_window';
